fix: widen synapse from_index u16→u32 for GPU mirror
neat-core's SynapseData.from_index is u16 (≤ u16::MAX), but the GPU
mirror SynapseGpu and WGSL shader use u32 (WGSL has no u16 type). The
direct field assignment caused E0308 mismatched types, breaking the
release build (RUSTFLAGS="-D warnings"). Widen explicitly with
u32::from(), matching the existing num_synapses conversion pattern.

Surfaced by the failing "Unit tests + coverage" check on
stSoftwareAU/NEAT-AI-Examples PR #606, which builds rust_scorer.

Vibe-Coder-Run-Id: vibe-mqqdj58s-b50a27

+++ rust_scorer/src/gpu/forward_mse_batched.rs
@@ -225,7 +225,9 @@ pub fn build_batched_network_data(
         for s in &net.synapses {
             synapses.push(SynapseGpu {
                 weight: s.weight,
-                from_index: s.from_index,
+                // neat-core stores `from_index` as u16 (≤ u16::MAX); the GPU
+                // mirror and WGSL shader use u32, so widen explicitly.
+                from_index: u32::from(s.from_index),
             });
         }
